Rename probe folder
