+++ internal/pkg/instrumentation/bpf/go.opentelemetry.io/otel/traceglobal/bpf_bpfel_arm64.go
