+++ internal/pkg/instrumentation/bpf/go.opentelemetry.io/otel/traceglobal/bpf_bpfel_x86.go
